+++ data-raw/packages.R
@@ -32,6 +32,7 @@ paletteer_packages <- tibble::tribble(
   "MetBrewer",          "BlakeRMills/MetBrewer",            "0.2.0",       TRUE,  "0.2.0",
   "Manu",               "G-Thomson/Manu",                   "0.0.2",       FALSE, NA,
   "MapPalettes",        "disarm-platform/MapPalettes",      "0.0.2",       FALSE, NA,
+  "musculusColors",     "dawnbarlow/musculusColors",        "0.1.0",       FALSE, NA,
   "miscpalettes",       "EmilHvitfeldt/miscpalettes",       "0.0.0.9000",  FALSE, NA,
   "nationalparkcolors", "katiejolly/nationalparkcolors",    "0.1.0",       FALSE, NA,
   "nbapalettes",        "murrayjw/nbapalettes",             "0.1.0.9000",  TRUE,  "0.1.0",