+++ NEWS.md
@@ -4,6 +4,8 @@
 * Added 31 Steven Universe palettes from tvthemes package.
 * 11 palettes have been added from the feathers package. (#116)
 * 56 palettes have been added from the MetBrewer package. (#119)
+* 7 palettes have been added from the musculusColors package. (#120)
+
 # paletteer 1.4.1
 
 * Fixed documentation to be HTML5 friendly.